added constants
