+++ src/constants/Constants.js
@@ -1,0 +1,23 @@
+const BASE_NBA_DATA_URL = "http://data.nba.com";
+const BASE_NBA_DATA_SCOREBOARD_URL = BASE_NBA_DATA_URL.concat("/data/5s/json/cms/noseason/scoreboard/");
+
+const DEFAULT_DATE_FORMAT = "YYYYMMDD";
+const DEFAULT_TIMEZONE = "America/New_York";
+
+const TRANSLATED_NBA_DATE_TIME_FORMAT = "YYYYMMDDHHmm";
+const TRANSLATED_DATE_FORMAT = "LLL";
+
+const TRANSLATED_GAME_STATUS_MAP = {
+  1: 'PREGAME',
+  2: 'LIVE',
+  3: 'FINAL'
+};
+
+module.exports = {
+  BASE_NBA_DATA_SCOREBOARD_URL: BASE_NBA_DATA_SCOREBOARD_URL,
+  DEFAULT_DATE_FORMAT: DEFAULT_DATE_FORMAT,
+  DEFAULT_TIMEZONE: DEFAULT_TIMEZONE,
+  TRANSLATED_NBA_DATE_TIME_FORMAT: TRANSLATED_NBA_DATE_TIME_FORMAT,
+  TRANSLATED_DATE_FORMAT: TRANSLATED_DATE_FORMAT,
+  TRANSLATED_GAME_STATUS_MAP: TRANSLATED_GAME_STATUS_MAP,
+};
